@@ -1,9 +1,9 @@
-module github.com/avito-tech/go-transaction-manager/v2/drivers/goredis8/v2
+module github.com/avito-tech/go-transaction-manager/drivers/goredis8/v2
 
-go 1.13
+go 1.14
 
 require (
-	github.com/avito-tech/go-transaction-manager/v2 v2.0.0-rc4
+	github.com/avito-tech/go-transaction-manager/trm/v2 v2.0.0-rc5
 	github.com/go-redis/redis/v8 v8.11.5
 	github.com/go-redis/redismock/v8 v8.11.5
 	github.com/stretchr/testify v1.8.2
