@@ -1,9 +1,9 @@
-module github.com/avito-tech/go-transaction-manager/v2/drivers/pgxv4/v2
+module github.com/avito-tech/go-transaction-manager/drivers/pgxv4/v2
 
-go 1.13
+go 1.16
 
 require (
-	github.com/avito-tech/go-transaction-manager/v2 v2.0.0-rc4
+	github.com/avito-tech/go-transaction-manager/trm/v2 v2.0.0-rc5
 	github.com/jackc/pgconn v1.14.1
 	github.com/jackc/pgx/v4 v4.18.1
 	github.com/pashagolub/pgxmock v1.8.0
