@@ -1,10 +1,27 @@
-module github.com/avito-tech/go-transaction-manager/v2/drivers/pgxv5/v2
+module github.com/avito-tech/go-transaction-manager/drivers/pgxv5/v2
 
-go 1.13
+go 1.19
 
 require (
-	github.com/avito-tech/go-transaction-manager/v2 v2.0.0-rc4
+	github.com/avito-tech/go-transaction-manager/trm/v2 v2.0.0-rc5
 	github.com/jackc/pgx/v5 v5.5.1
 	github.com/pashagolub/pgxmock/v2 v2.12.0
 	github.com/stretchr/testify v1.8.2
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/golang/mock v1.6.0 // indirect
+	github.com/jackc/pgpassfile v1.0.0 // indirect
+	github.com/jackc/pgservicefile v0.0.0-20221227161230-091c0ba34f0a // indirect
+	github.com/jackc/puddle/v2 v2.2.1 // indirect
+	github.com/kr/text v0.2.0 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/rogpeppe/go-internal v1.12.0 // indirect
+	go.uber.org/atomic v1.7.0 // indirect
+	go.uber.org/multierr v1.9.0 // indirect
+	golang.org/x/crypto v0.9.0 // indirect
+	golang.org/x/sync v0.1.0 // indirect
+	golang.org/x/text v0.9.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
