@@ -1,6 +1,3 @@
-//go:build go1.16
-// +build go1.16
-
 package gorm
 
 import (
@@ -8,8 +5,8 @@ import (
 
 	"gorm.io/gorm"
 
-	trm "github.com/avito-tech/go-transaction-manager/v2"
-	trmcontext "github.com/avito-tech/go-transaction-manager/v2/context"
+	"github.com/avito-tech/go-transaction-manager/trm/v2"
+	trmcontext "github.com/avito-tech/go-transaction-manager/trm/v2/context"
 )
 
 // DefaultCtxGetter is the CtxGetter with settings.DefaultCtxKey.
