@@ -3,7 +3,7 @@ package mongo
 import (
 	"context"
 
-	trm "github.com/avito-tech/go-transaction-manager/v2"
+	trm "github.com/avito-tech/go-transaction-manager/trm/v2"
 )
 
 // NewDefaultFactory creates default trm.Transaction(mongo.Session).
