@@ -1,12 +1,9 @@
-//go:build go1.16
-// +build go1.16
-
 package pgxv4
 
 import (
 	"context"
 
-	trm "github.com/avito-tech/go-transaction-manager/v2"
+	"github.com/avito-tech/go-transaction-manager/trm/v2"
 )
 
 // NewDefaultFactory creates default trm.Transaction(pgx.Tx).
