@@ -1,6 +1,3 @@
-//go:build go1.16
-// +build go1.16
-
 package pgxv4
 
 import (
@@ -9,7 +6,7 @@ import (
 
 	"github.com/jackc/pgx/v4"
 
-	trm "github.com/avito-tech/go-transaction-manager/v2"
+	"github.com/avito-tech/go-transaction-manager/trm/v2"
 )
 
 // Transaction is trm.Transaction for pgx.Tx.
