@@ -1,6 +1,3 @@
-//go:build go1.16
-// +build go1.16
-
 package pgxv4
 
 import (
@@ -15,14 +12,13 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	trm "github.com/avito-tech/go-transaction-manager/v2"
-	"github.com/avito-tech/go-transaction-manager/v2/mock"
+	"github.com/avito-tech/go-transaction-manager/trm/v2/mock"
 
-	trmcontext "github.com/avito-tech/go-transaction-manager/v2/context"
+	trmcontext "github.com/avito-tech/go-transaction-manager/trm/v2/context"
 
-	"github.com/avito-tech/go-transaction-manager/v2/settings"
-
-	"github.com/avito-tech/go-transaction-manager/v2/manager"
+	"github.com/avito-tech/go-transaction-manager/trm/v2"
+	"github.com/avito-tech/go-transaction-manager/trm/v2/manager"
+	"github.com/avito-tech/go-transaction-manager/trm/v2/settings"
 )
 
 func TestTransaction(t *testing.T) {
