@@ -1,6 +1,3 @@
-//go:build go1.19
-// +build go1.19
-
 package pgxv5
 
 import (
@@ -9,8 +6,8 @@ import (
 	"github.com/jackc/pgx/v5"
 	"github.com/stretchr/testify/assert"
 
-	trm "github.com/avito-tech/go-transaction-manager/v2"
-	"github.com/avito-tech/go-transaction-manager/v2/settings"
+	"github.com/avito-tech/go-transaction-manager/trm/v2"
+	"github.com/avito-tech/go-transaction-manager/trm/v2/settings"
 )
 
 func TestSettings_EnrichBy(t *testing.T) {
