@@ -6,11 +6,12 @@ import (
 
 	_ "github.com/mattn/go-sqlite3"
 
-	trm "github.com/avito-tech/go-transaction-manager/v2"
-	trmcontext "github.com/avito-tech/go-transaction-manager/v2/context"
-	trmsqlx "github.com/avito-tech/go-transaction-manager/v2/drivers/sqlx/v2"
-	"github.com/avito-tech/go-transaction-manager/v2/manager"
-	"github.com/avito-tech/go-transaction-manager/v2/settings"
+	"github.com/avito-tech/go-transaction-manager/trm/v2"
+	trmcontext "github.com/avito-tech/go-transaction-manager/trm/v2/context"
+	"github.com/avito-tech/go-transaction-manager/trm/v2/manager"
+	"github.com/avito-tech/go-transaction-manager/trm/v2/settings"
+
+	trmsqlx "github.com/avito-tech/go-transaction-manager/drivers/sqlx/v2"
 )
 
 // Example demonstrates a work of manager.ChainedMW.
