@@ -7,8 +7,8 @@ import (
 	"github.com/jmoiron/sqlx"
 	_ "github.com/mattn/go-sqlite3"
 
-	trmsqlx "github.com/avito-tech/go-transaction-manager/v2/drivers/sqlx/v2"
-	"github.com/avito-tech/go-transaction-manager/v2/manager"
+	trmsqlx "github.com/avito-tech/go-transaction-manager/drivers/sqlx/v2"
+	"github.com/avito-tech/go-transaction-manager/trm/v2/manager"
 )
 
 // Example demonstrates the implementation of the Repository pattern by trm.Manager.
