@@ -5,9 +5,9 @@ import (
 
 	"github.com/jmoiron/sqlx"
 
-	trmsql "github.com/avito-tech/go-transaction-manager/v2/drivers/sql/v2"
+	trm "github.com/avito-tech/go-transaction-manager/trm/v2"
 
-	trm "github.com/avito-tech/go-transaction-manager/v2"
+	trmsql "github.com/avito-tech/go-transaction-manager/drivers/sql/v2"
 )
 
 // NewDefaultFactory creates default trm.Transaction(sqlx.Tx).
