@@ -9,9 +9,8 @@ import (
 	"github.com/jmoiron/sqlx"
 	"go.uber.org/multierr"
 
-	trmsql "github.com/avito-tech/go-transaction-manager/v2/drivers/sql/v2"
-
-	trm "github.com/avito-tech/go-transaction-manager/v2"
+	trmsql "github.com/avito-tech/go-transaction-manager/drivers/sql/v2"
+	"github.com/avito-tech/go-transaction-manager/trm/v2"
 )
 
 // Transaction is trm.Transaction for sqlx.Tx.
